requirements.txt
colorful==0.5.4
PyMySQL==0.10.1
six==1.15.0
ConfigParser==5.0.1
prettyprinter~=0.18.0
lxml~=4.6.2
setuptools~=51.1.0.post20201221
requests~=2.25.1
openpyxl~=3.0.5
hyper~=0.7.0
PyExecJS~=1.5.1
hpack~=2.2.0
certifi~=2020.12.5
chardet~=4.0.0
urllib3~=1.26.2
jdcal~=1.4.1
Pillow~=8.0.1
Pygments~=2.7.3
idna~=2.10
zhon~=1.1.5
numpy~=1.19.3
dict_hash~=1.0.8
cffi~=1.14.4
wheel~=0.36.2
keyring~=21.8.0
rsa~=4.6
pyasn1~=0.4.8
gast~=0.3.3
h5py
pytz~=2020.5
PyYAML~=5.3.1
Naked~=0.1.31
shellescape~=3.8.1
cachetools~=4.2.0
pandas~=1.1.5
python-dateutil~=2.8.1
Markdown~=3.3.3
oauthlib~=3.1.0
Werkzeug~=1.0.1
pycparser~=2.20
flatbuffers~=1.12
wrapt~=1.12.1
astunparse~=1.6.3
termcolor~=1.1.0
crypto~=1.4.1
matplotlib~=3.3.4
opencv-python~=4.5.1.48
websocket-client~=1.0.1